Create Compare Egg & Feed Prices
Separating a pull from FRED to pull and compare the prices of the Egg Prices (FRED Series APU0000708111) to the Price of Fowl Feed (FRED Series PCU3111193111191). Including a time lag comparison between the change of the two series.

Ultimately when we review this information, at least for the time frame the code is reviewing (2005 - 2025) there is no useful correlation. All sequences are less than 0.10.